+++ src/Stack/Options/SetupParser.hs
@@ -20,7 +20,7 @@ setupOptsParser = SetupCmdOpts
   <$> OA.optional (OA.argument readVersion
         (  OA.metavar "GHC_VERSION"
         <> OA.help "Version of GHC to install, e.g. 9.6.4. (default: install \
-                   \the version implied by the resolver)"
+                   \the version implied by the snapshot)"
         ))
   <*> OA.boolFlags False
         "reinstall"