+++ src/Stack/Types/Project.hs
@@ -27,17 +27,17 @@ data Project = Project
     -- of the snapshot.
   , flagsByPkg :: !(Map PackageName (Map FlagName Bool))
     -- ^ Flags to be applied on top of the snapshot flags.
-  , resolver :: !RawSnapshotLocation
+  , snapshot :: !RawSnapshotLocation
     -- ^ How we resolve which @Snapshot@ to use
   , compiler :: !(Maybe WantedCompiler)
-    -- ^ Override the compiler in 'projectResolver'
+    -- ^ Override the compiler in 'snapshot'
   , extraPackageDBs :: ![FilePath]
   , curator :: !(Maybe Curator)
     -- ^ Extra configuration intended exclusively for usage by the curator tool.
     -- In other words, this is /not/ part of the documented and exposed Stack
     -- API. SUBJECT TO CHANGE.
   , dropPackages :: !(Set PackageName)
-    -- ^ Packages to drop from the 'projectResolver'.
+    -- ^ Packages to drop from the 'snapshot'.
   }
   deriving Show
 
@@ -54,7 +54,7 @@ instance ToJSON Project where
       | not (Map.null project.flagsByPkg)
       ]
     , ["packages" .= project.packages]
-    , ["snapshot" .= project.resolver]
+    , ["snapshot" .= project.snapshot]
     , maybe [] (\c -> ["curator" .= c]) project.curator
     , [ "drop-packages" .= Set.map CabalString project.dropPackages
       | not (Set.null project.dropPackages)