+++ tests/integration/tests/3631-build-http2/Main.hs
@@ -2,5 +2,5 @@ import StackTest
 
 main :: IO ()
 main = do
-  stack ["build", defaultResolverArg, "--dry-run", "http2"]
-  stack ["build", defaultResolverArg, "http2"]
+  stack ["build", defaultSnapshotArg, "--dry-run", "http2"]
+  stack ["build", defaultSnapshotArg, "http2"]