+++ tests/integration/tests/384-local-deps/Main.hs
@@ -3,5 +3,5 @@ import StackTest
 main :: IO ()
 main = do
     removeFileIgnore "stack.yaml"
-    stack ["init", defaultResolverArg]
+    stack ["init", defaultSnapshotArg]
     stack ["test"]