+++ tests/integration/tests/4453-detailed/Main.hs
@@ -4,4 +4,4 @@ import System.Directory
 main :: IO ()
 main = do
     removeDirIgnore ".stack-work"
-    stack [defaultResolverArg, "test"]
+    stack [defaultSnapshotArg, "test"]